requirements.txt
websocket-client
rel
pyautogui